requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
drf-excel==2.4.0
et-xmlfile==1.1.0
openpyxl==3.1.2
psycopg==3.1.12
psycopg-binary==3.1.12
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
unicodecsv==0.14.1
